% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_raw_datasets.R
\name{download_raw_datasets}
\alias{download_raw_datasets}
\title{Cette fonction permet de télécharger les données csv de data.gouv.fr.}
\usage{
download_raw_datasets(years, path = getwd(), quiet = TRUE)
}
\arguments{
\item{years}{Vecteur numérique des années à télécharger}

\item{path}{Vecteur de caractères des noms de chemin complets ; la valeur par défaut correspond au répertoire de travail, getwd()}

\item{quiet}{Si TRUE, supprime les messages et la barre de progression; la valeur par défaut est TRUE.}
}
\value{
Un code entier (invisible), 0 pour le succès et non zéro pour l'échec.
}
\description{
Cette fonction permet de télécharger les données csv de data.gouv.fr.
}
\examples{
\dontrun{
download_raw_datasets(2018, tempdir())
download_raw_datasets(2016:2018, tempdir())
}
}
