% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datalist.R
\name{get_datalist}
\alias{get_datalist}
\title{Récupère la liste des données disponibles sur data.gouv.fr.}
\usage{
get_datalist(years = NA_integer_)
}
\arguments{
\item{years}{Vecteur numérique des années à télécharger}
}
\value{
Renvoie un data.frame correspondant à la liste de l'ensemble des données disponibles.
}
\description{
Récupère la liste des données disponibles sur data.gouv.fr.
}
\examples{
get_datalist()
get_datalist(2018)
get_datalist(2016:2018)
}
