% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicule.R
\docType{data}
\name{vehicule}
\alias{vehicule}
\title{Véhicules impliqués dans l'accident}
\format{
Un data.frame avec 136021 observations et 13 variables:
\describe{
\item{Num_Acc}{Identifiant de l’accident identique à celui du fichier "rubrique CARACTERISTIQUES" repris pour chacun des usagers décrits impliqués dans l’accident}
\item{num_veh}{Identifiant du véhicule repris pour chacun des usagers occupant ce véhicule (y compris les piétons qui sont rattachés aux véhicules qui les ont heurtés) – Code alphanumérique}
\item{senc}{Sens de circulation (1 = PK ou PR ou numéro d’adresse postale croissant, 2 = PK ou PR ou numéro d’adresse postale décroissant)}
\item{catv}{Catégorie du véhicule (01 = Bicyclette, 02 = Cyclomoteur <50cm3, 03 = Voiturette (Quadricycle à moteur carrossé) (anciennement "voiturette ou tricycle à moteur"), 04 = Référence plus utilisée depuis 2006 (scooter immatriculé), 05 = Référence plus utilisée depuis 2006 (motocyclette), 06 = Référence plus utilisée depuis 2006 (side-car), 07 = VL seul, 08 = Catégorie plus utilisée (VL + caravane), 09 = Catégorie plus utilisée (VL + remorque), 10 = VU seul 1,5T <= PTAC <= 3,5T avec ou sans remorque (anciennement VU seul 1,5T <= PTAC <= 3,5T), 11 = Référence plus utilisée depuis 2006 (VU (10) + caravane), 12 = Référence plus utilisée depuis 2006 (VU (10) + remorque), 13 = PL seul 3,5T <PTCA <= 7,5T, 14 = PL seul > 7,5T, 15 = PL > 3,5T + remorque, 16 = Tracteur routier seul, 17 = Tracteur routier + semi-remorque, 18 = Référence plus utilisée depuis 2006 (transport en commun), 19 = Référence plus utilisée depuis 2006 (tramway), 20 = Engin spécial, 21 = Tracteur agricole, 30 = Scooter < 50 cm3, 31 = Motocyclette > 50 cm3 et <= 125 cm3, 32 = Scooter > 50 cm3 et <= 125 cm3, 33 = Motocyclette > 125 cm3, 34 = Scooter > 125 cm3, 35 = Quad léger <= 50 cm3 (Quadricycle à moteur non carrossé), 36 = Quad lourd > 50 cm3 (Quadricycle à moteur non carrossé), 37 = Autobus, 38 = Autocar, 39 = Train, 40 = Tramway, 99 = Autre véhicule (dont piéton en roller ou en trottinette à partir de l’année 2018 requalifié en "engin de déplacement personnel"))}
\item{obs}{Obstacle fixe heurté (1 = Véhicule en stationnement, 2 = Arbre, 3 = Glissière métallique, 4 = Glissière béton, 5 = Autre glissière, 6 = Bâtiment, mur, pile de pont, 7 = Support de signalisation verticale ou poste d’appel d’urgence, 8 = Poteau, 9 = Mobilier urbain, 10 = Parapet, 11 = Ilot, refuge, borne haute, 12 = Bordure de trottoir, 13 = Fossé, talus, paroi rocheuse, 14 = Autre obstacle fixe sur chaussée, 15 = Autre obstacle fixe sur trottoir ou accotement, 16 = Sortie de chaussée sans obstacle)}
\item{obsm}{Obstacle mobile heurté (1 – Piéton, 2 – Véhicule, 4 – Véhicule sur rail, 5 – Animal domestique, 6 – Animal sauvage, 9 – Autre )}
\item{choc}{Point de choc initial (1 - Avant, 2 – Avant droit, 3 – Avant gauche, 4 – Arrière, 5 – Arrière droit, 6 – Arrière gauche, 7 – Côté droit, 8 – Côté gauche, 9 – Chocs multiples (tonneaux))}
\item{manv}{Manœuvre principale avant l’accident (1 – Sans changement de direction, 2 – Même sens, même file, 3 – Entre 2 files, 4 – En marche arrière, 5 – A contresens, 6 – En franchissant le terre-plein central, 7 – Dans le couloir bus, dans le même sens, 8 – Dans le couloir bus, dans le sens inverse, 9 – En s’insérant, 10 – En faisant demi-tour sur la chaussée, 11 – Changeant de file - A gauche, 12 – Changeant de file - A droite, 13 – Déporté - A gauche, 14 – Déporté - A droite, 15 – Tournant - A gauche, 16 – Tournant - A droite, 17 – Dépassant - A gauche, 18 – Dépassant - A droite, 19 – Traversant la chaussée, 20 – Manœuvre de stationnement, 21 – Manœuvre d’évitement, 22 – Ouverture de porte, 23 – Arrêté (hors stationnement), 24 – En stationnement (avec occupants))}
\item{occutc}{Nombre d’occupants dans le transport en commun}
}
}
\source{
\url{https://www.data.gouv.fr/fr/datasets/base-de-donnees-accidents-corporels-de-la-circulation/#_}
}
\usage{
vehicule
}
\description{
Base de données contenant des informations sur les véhicules impliqués dans l'accident.
}
\keyword{datasets}
