% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concat_data.R
\name{get_concat_data}
\alias{get_concat_data}
\title{Renvoie la table unique des données des années specifiées.}
\usage{
get_concat_data(years = NA_integer_)
}
\arguments{
\item{years}{Vecteur numérique des années à télécharger; la valeur par défaut est NA_integer_}
}
\value{
Un data.frame correspondant la jointure des différentes tables
}
\description{
Renvoie la table unique des données des années specifiées.
}
\examples{
\dontrun{
get_concat_data(2018)
}
}
