% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Charge l'ensemble des données des années spécifiées.}
\usage{
get_data(years)
}
\arguments{
\item{years}{Vecteur numérique des années à télécharger}
}
\value{
Renvoie une liste contenant les tables caractéristique, lieux, usagers et véhicules.
}
\description{
Charge l'ensemble des données des années spécifiées.
}
\examples{
get_data(2018)
}
